Go versionless for v2.0
DOESN'T WORK WITH TOOLING YET.
This wasn't working in 6.x due to https://github.com/dotnet/Nerdbank.GitVersioning/issues/828 which HAS been fixed, but is not working on 7.0.300 7.0.300-preview.23179.2 due to https://github.com/dotnet/maui/issues/14524 which has NOT been fixed yet. Committing now as it should work in the future with a simple SDK update.

We won't increment the version number past 2.0, and will let the automatic versioning handle it while also publishing a version to the app store for every commit thats pushed to the main branch instead of relying on version tags.

Considered going back to 1.0 for this, but since version numbers are imaginary, lets call it 2.0. Or maybe AntDesign was 1.x and MudBlazor is 2.x, who knows.