Handle main window deactivation from index
For some reason injecting the db context in two different places makes the change tracking not work right. No idea why so I'm just moving it to the index for now.

Also changes some controls to be Immediate so closing the window while they're in focus won't lose the changes.