Make Safety Plan always available
The idea was to make it discoverable but out of the way if you don't need it, but if you make your own categories it will disappear. Being locked into a category is not what I want. I can also imagine a situation when one person demostrate the app to another and they are confused by the lack of the safety plan.

Also removes a DB call :)