Make all default categories writeable except Mood
I considered removing some because there were just too many filling up the list, but is the best of both worlds, giving people easy access to default ones, but also allowing them to delete (or change the name of) the ones they want to.

Overall mood is required for the Mood Grid.