Fix Import/export by using JSON instead of CSV
Handles three major issues:
- null couldn't be distinguished from empty string
- object references weren't preserved
- hard to save properties that don't belong in a list

I kind of liked the accessibility of the CSV files but it just wasn't working out with the way EF Core handles one-to-many relationships.

Importing over the same dataset you exported doesn't seem to be working yet, but we can fix that later.

This is the first officially working iteration of the import/export feature and so it's been enabled for production.