Share export instead of writing directly to avoid permission dependency
"modify or delete the contents of your USB storage" is an ugly permission and I don't want to see that when you go to download it. This gets rid of that and could be more user friendly anyway!

I'll try and drop the READ_EXTERNAL_STORAGE permission next. Maybe by making the app a share target and the default app for .journalapp files.