Go back to using long-lived DB context in Index and temporarily disable import
We switched over to fix import not updating anything but I fixed that now, and because it seemed like good practice but it was difficult to work around and was causing issues like points edited in the index not saved.

Now we simply won't save the keys in the export so when you import again it won't get mixed up. Has the bonus of hugely reducing file size.

Disabling import for now has a great side effect of dropping the ugly "read external storage" permission. When we re-enable import I was it to work via file activation instead so we don't have to have any permissions at all.