Revert "WIP of notfication reminders"
This reverts commit 451bef088856dcf419423db0431ca8c03961fa65.

I really don't like the extra permissions which felt like hacks and I just couldn't get it working right anyway.

I've also always felt like it would be better to delegate this to actual reminder apps. But having it in the app could be easier for the user as they would be able to click the notification to open the app and it could build user retention by making it easier to set the notifications in the first place.

May give it another shot in the future if I recieve more feedback about it.