Remove Trends name customization and export scaffolding
I want to add these features but they're too cumbersome so I'll put them on hold indefinitely until I get feedback on it.

You can honestly just take a screenshot of the app and use the built-in scroll to get the same effect that I was going for. Alternative was to create a separate document specially made for Letter format but that's a whole lot of extra work for something that might be used very rarely.