Revert "Remove unused target"
This reverts commit d6451285be0ac89735cad9a78957ecf4a4c145d3.

Totally overlooked the part that this is necssary for the tests. I removed it because the Plugin.Maui.AppRating did not support it, but that's changed as of https://github.com/FabriBertani/Plugin.Maui.AppRating/commit/72b99a42f1ab25c09168dfc9b3c21f208490bbc6 and I just have to wait for the next release.