Remove AM/PM example meds and replace with two other defaults
Having so many defaults was too overwhelming and actually slowed the initialization down (not well optimized), so I've been cutting back. The AM/PM examples are poor because they do not allow the dose feature and don't give a proper representation to go off. This should help with onboarding even if users question why there are two specific pre-existing meds.